Refactor utility classes/methods
Move most of the non-satellite/signal utilities out of GPSTestUtil and rename it to SatelliteUtils
